fix(backend): make the migrations survive MariaDB 5.5, and cast ttl_days
Deploying found three things local sqlite could not.

The shared server runs MariaDB 5.5.68 (zachs_app_base.md §11.1), which has
no JSON column type: shared_plays.data is longText now. The repo had also
drifted from production on the same point — tactics.data was fixed to
longText on the server when it was first deployed and the correction never
came back, so a fresh deploy from this repo would have failed on a table
that has worked for months. Both are longText here now.

MariaDB 5.5 indexes utf8mb4 at four bytes a character, so a default
varchar(255) key blows its 767-byte index limit — Laravel's own cache table
keys on `key`, so it simply would not create. AppServiceProvider now sets
Builder::defaultStringLength(191).

And ttl_days: the app posts multipart, so it arrives as the string "7" and
Carbon::addDays() rejects a string outright with a TypeError. Validation
says integer but does not cast. Every share with an explicit TTL was a 500;
share_link_service.dart does send the field.

Deployed and verified against production: POST /api/v1/share returns a slug,
the API and /p/{slug} both serve it, an unknown slug 404s, shares:prune runs
clean, and a scheduler cron now ticks it every minute.

+++ backend/app/Http/Controllers/SharedPlayController.php
@@ -58,8 +58,11 @@ class SharedPlayController extends Controller
             'title' => $validated['title'] ?? null,
             'sport_type' => $validated['sport_type'],
             'data' => $data,
+            // (int) is not decoration: the app posts multipart, so ttl_days
+            // arrives as the string "7" and Carbon::addDays() rejects a string
+            // outright. Validation says integer but does not cast.
             'expires_at' => Carbon::now()->addDays(
-                $validated['ttl_days'] ?? self::DEFAULT_TTL_DAYS
+                (int) ($validated['ttl_days'] ?? self::DEFAULT_TTL_DAYS)
             ),
         ]);
